+++ src/Presentation/SmartStore.Web/Infrastructure/Routes/MapLegacyRoutesAttribute.cs
@@ -6,20 +6,46 @@ using System.Web;
 using System.Web.Mvc;
 using System.Web.Routing;
 using SmartStore.Core.Data;
+using SmartStore.Core.Infrastructure;
 using SmartStore.Utilities;
+using SmartStore.Web.Framework;
 
 namespace SmartStore.Web.Infrastructure
 {
-	
+	public class LegacyRoutesStarter : IStartupTask
+	{
+		internal const string MediaPattern = @"/media/(?<dir>thumbs|uploaded)/(?<path>.*?)$";
+
+		public void Execute()
+		{
+			if (CommonHelper.GetAppSetting<bool>("sm:EnableLegacyRoutesMapping"))
+			{
+				SmartUrlRoutingModule.RegisterRoutablePath(MediaPattern, "GET");
+			}
+		}
+
+		public int Order
+		{
+			get { return 0; }
+		}
+	}
+
 	public sealed class MapLegacyRoutesAttribute : ActionFilterAttribute
 	{
+		private static readonly bool s_enabled = CommonHelper.GetAppSetting<bool>("sm:EnableLegacyRoutesMapping");
 		private static readonly IList<MappedRoute> s_mappedRoutes = new List<MappedRoute>();
 
 		static MapLegacyRoutesAttribute()
 		{
-			AddLegacyRoute(@"/productreviews/(?<id>\d+)$", "/product/reviews/${id}", "GET");
-			AddLegacyRoute("/sitemapseo$", "/sitemap.xml", "GET");
-			AddLegacyRoute("/config$", "/settings", "GET");
+			if (s_enabled)
+			{
+				// /Media/(Thumbs|Uploaded)/{*path} > /Media/Default/(Thumbs|Uploaded)/{*path}
+				AddLegacyRoute(LegacyRoutesStarter.MediaPattern, "/media/" + DataSettings.Current.TenantName + "/${dir}/${path}", "GET");
+
+				AddLegacyRoute(@"/productreviews/(?<id>\d+)$", "/product/reviews/${id}", "GET");
+				AddLegacyRoute("/sitemapseo$", "/sitemap.xml", "GET");
+				AddLegacyRoute("/config$", "/settings", "GET");
+			}
 		}
 
 		private static void AddLegacyRoute(string path, string newPath, string verb = ".*")
@@ -32,13 +58,13 @@ namespace SmartStore.Web.Infrastructure
 
 		public override void OnActionExecuted(ActionExecutedContext filterContext)
 		{
+			if (!s_enabled)
+				return;
+
 			if (!DataSettings.DatabaseIsInstalled())
 				return;
 
 			if (filterContext.IsChildAction)
-				return;
-
-			if (!CommonHelper.GetAppSetting<bool>("sm:EnableLegacyRoutesMapping"))
 				return;
 
 			var mappedPath = GetMappedPath(filterContext.HttpContext.Request);